Update driver controller and service

+++ src/main/java/assignment2/model/Driver.java
@@ -65,7 +65,7 @@ public class Driver {
 	public void setPhone(String phone) {
 		this.phone = phone;
 	}
-	public void setRating(int rating) {
+	public void setRating(double rating) {
 		this.rating = rating;
 	}
 }
